+++ packages/gatsby-theme-docs/src/layouts/internals/sidebar.utils.js
@@ -1,17 +1,23 @@
-export const getItemDescendants = (level, index, ancestorsMap) => {
+export const getItemDescendants = (
+  subChapterId,
+  level,
+  index,
+  ancestorsMap
+) => {
   const chapterId = `${level}-${index}`;
   if (level === 1) {
-    return [chapterId];
+    return [subChapterId];
   } else {
     const descendantsArray = ancestorsMap.filter((element) =>
       element.includes(chapterId)
     );
-    return [].concat(...descendantsArray);
+    return (chapterId.includes('#') ? [subChapterId] : []).concat(
+      ...descendantsArray
+    );
   }
 };
 
-export const getItemAncestors = (level, index, ancestorsMap) => {
-  const chapterId = `${level}-${index}`;
+export const getItemAncestors = (chapterId, ancestorsMap) => {
   const ancestorsArray = ancestorsMap.filter(
     (element) => element.includes(chapterId) && element.indexOf(chapterId) > 0
   );