+++ websites/api-docs-smoke-test/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @commercetools-website/api-docs-smoke-test
 
+## 8.4.9
+
+### Patch Changes
+
+- Updated dependencies [[`599a11e`](https://github.com/commercetools/commercetools-docs-kit/commit/599a11e7d10526ecae1849374ca9e8b02af9d3db), [`d123efa`](https://github.com/commercetools/commercetools-docs-kit/commit/d123efa33e4703447c0192461352d244a7fe94f6)]:
+  - @commercetools-docs/gatsby-theme-docs@23.0.0
+  - @commercetools-docs/gatsby-theme-api-docs@23.0.0
+
 ## 8.4.8
 
 ### Patch Changes